[agile] Capture: Book cost_center/accounts link + flagged dynamic combo codegen gap
