+++ doc/agile/product_backlog/inbox/book_cost_center_should_reference_accounts.org
@@ -1,0 +1,43 @@
+:PROPERTIES:
+:ID: 847E30FE-FA31-4940-B332-3A56B27B1334
+:END:
+#+title: Book cost_center should reference a real accounts entity
+#+description: Book.cost_center is currently a free-form nullable string with no validation or link to any ledger/account structure.
+#+type: capture
+#+level: cross
+#+filetags: :refdata:book:accounts:data-modeling:
+#+created: 2026-07-06
+#+updated: 2026-07-06
+
+This page is a [[id:671F18E4-E09C-4B3B-BD24-D33DF8AE38A6][capture]] in the *inbox* bucket of the product backlog — a pre-sprint idea, not yet pulled into a sprint as a story.
+
+* What
+
+=Book.cost_center= (=projects/ores.refdata/modeling/ores.refdata.book.org:145-152=) is
+declared as a plain nullable =std::string= — "Internal finance code for
+P&L attribution" — with no validation, no foreign key, and no link to
+any ledger/chart-of-accounts structure. There is currently no
+=ores.accounts= (or equivalent) component in the codebase at all.
+Investigate what a "cost centre" should really be: a reference to a
+chart-of-accounts / cost-centre master entity, likely owned by a new or
+existing accounts-domain component, and reconcile Book's field
+accordingly (soft-FK combo, validation, or a dedicated entity).
+
+* Why
+
+A free-text cost centre invites drift (typos, inconsistent codes,
+no way to rename/reorganise centrally) and blocks any reporting or
+validation that depends on cost centres being a closed, well-known set.
+Matching it to a real accounts entity would let the Qt layer render it
+as a validated combo (same pattern as other soft-FK fields) instead of
+an unchecked string.
+
+* References
+
+- =projects/ores.refdata/modeling/ores.refdata.book.org= — current
+  =cost_center= field definition (line 145) and Qt field/column mapping
+  (lines 280-292).
+
+* See also
+
+- [[id:04BE9B6F-B43C-426B-96D1-AED89F88FA93][Story: Commission: book]] — where this was noticed.
