+++ projects/ores.qt/refdata/src/BookDetailDialog.cpp
@@ -21,7 +21,7 @@
 #include "ores.qt/ChangeReasonDialog.hpp"
 #include "ores.qt/IconUtils.hpp"
 #include "ores.qt/MessageBoxHelper.hpp"
-#include "ores.refdata/messaging/book_protocol.hpp"
+#include "ores.refdata.api/messaging/book_protocol.hpp"
 #include "ui_BookDetailDialog.h"
 #include <QFutureWatcher>
 #include <QMessageBox>
