+++ projects/ores.qt/refdata/src/ClientBookModel.cpp
@@ -21,7 +21,7 @@
 #include "ores.qt/ColorConstants.hpp"
 #include "ores.qt/ExceptionHelper.hpp"
 #include "ores.qt/RelativeTimeHelper.hpp"
-#include "ores.refdata/messaging/book_protocol.hpp"
+#include "ores.refdata.api/messaging/book_protocol.hpp"
 #include <QtConcurrent>
 #include <boost/uuid/uuid_io.hpp>
