--- projects/ores.qt/trading/src/BookHistoryDialog.cpp
@@ -1,139 +1,0 @@
-/* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
- *
- * Copyright (C) 2026 Marco Craveiro <marco.craveiro@gmail.com>
- *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation; either version 3 of the License, or (at your option) any later
- * version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc., 51
- * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- *
- */
-#include "ores.qt/BookHistoryDialog.hpp"
-#include "ores.qt/RelativeTimeHelper.hpp"
-#include "ores.qt/WidgetUtils.hpp"
-#include "ores.refdata.api/messaging/book_protocol.hpp"
-#include "ui_BookHistoryDialog.h"
-#include <QHeaderView>
-#include <boost/uuid/uuid_io.hpp>
-
-namespace ores::qt {
-
-using namespace ores::logging;
-
-BookHistoryDialog::BookHistoryDialog(const boost::uuids::uuid& id,
-                                     const QString& code,
-                                     ClientManager* clientManager,
-                                     QWidget* parent)
-    : HistoryDialogBase(parent)
-    , ui_(new Ui::BookHistoryDialog)
-    , id_(id)
-    , code_(code)
-    , clientManager_(clientManager) {
-
-    ui_->setupUi(this);
-    WidgetUtils::setupComboBoxes(this);
-
-    ui_->titleLabel->setText(QString("History for: %1").arg(code_));
-
-    ui_->versionListWidget->setColumnCount(5);
-    ui_->versionListWidget->setHorizontalHeaderLabels(
-        {"Version", "Recorded At", "Modified By", "Performed By", "Commentary"});
-
-    initializeHistoryUi({.versionList = ui_->versionListWidget,
-                         .changesTable = ui_->changesTableWidget,
-                         .titleLabel = ui_->titleLabel,
-                         .closeButton = ui_->closeButton});
-}
-
-BookHistoryDialog::~BookHistoryDialog() = default;
-
-void BookHistoryDialog::loadHistory() {
-    BOOST_LOG_SEV(lg(), debug) << "Loading history for book: " << code_.toStdString();
-    emit statusChanged(tr("Loading history..."));
-
-    refdata::messaging::get_book_history_request request;
-    request.id = boost::uuids::to_string(id_);
-
-    runHistoryRequest(clientManager_, std::move(request), [this](auto response) {
-        if (!response.success) {
-            BOOST_LOG_SEV(lg(), error) << "Response was not success.";
-            historyLoadFailed(QString::fromStdString(response.message));
-            return;
-        }
-        versions_ = std::move(response.books);
-        historyLoaded();
-    });
-}
-
-int BookHistoryDialog::historySize() const {
-    return static_cast<int>(versions_.size());
-}
-
-HistoryDialogBase::VersionRow BookHistoryDialog::versionRow(int index) const {
-    const auto& version = versions_[index];
-    return {.version = version.version,
-            .cells = {relative_time_helper::format(version.recorded_at),
-                      QString::fromStdString(version.modified_by),
-                      QString::fromStdString(version.performed_by),
-                      QString::fromStdString(version.change_commentary)}};
-}
-
-QString BookHistoryDialog::historyTitle() const {
-    return QString("History for: %1").arg(code_);
-}
-
-HistoryDialogBase::DiffResult BookHistoryDialog::calculateDiffAt(int current_index,
-                                                                 int previous_index) const {
-    const auto& current = versions_[current_index];
-    const auto& previous = versions_[previous_index];
-
-    DiffResult diffs;
-    checkString(diffs, "Name", current.name, previous.name);
-    checkString(diffs, "Ledger Currency", current.ledger_ccy, previous.ledger_ccy);
-    checkString(diffs, "GL Account Ref", current.gl_account_ref, previous.gl_account_ref);
-    checkString(diffs, "Cost Center", current.cost_center, previous.cost_center);
-    checkString(diffs, "Status", current.book_status, previous.book_status);
-
-    return diffs;
-}
-
-void BookHistoryDialog::displayFullDetails(int index) {
-    const auto& version = versions_[index];
-
-    ui_->nameValue->setText(QString::fromStdString(version.name));
-    ui_->ledgerCcyValue->setText(QString::fromStdString(version.ledger_ccy));
-    ui_->glAccountRefValue->setText(QString::fromStdString(version.gl_account_ref));
-    ui_->costCenterValue->setText(QString::fromStdString(version.cost_center));
-    ui_->bookStatusValue->setText(QString::fromStdString(version.book_status));
-    ui_->versionNumberValue->setText(QString::number(version.version));
-    ui_->modifiedByValue->setText(QString::fromStdString(version.modified_by));
-    ui_->recordedAtValue->setText(relative_time_helper::format(version.recorded_at));
-    ui_->changeCommentaryValue->setText(QString::fromStdString(version.change_commentary));
-}
-
-void BookHistoryDialog::openVersionAt(int index) {
-    const auto& version = versions_[index];
-    BOOST_LOG_SEV(lg(), info) << "Opening book version " << version.version << " in read-only mode";
-    emit openVersionRequested(version, version.version);
-}
-
-void BookHistoryDialog::revertToVersionAt(int index) {
-    // The base has already confirmed with the user; revert TO the
-    // selected version.
-    const auto& selected = versions_[index];
-
-    BOOST_LOG_SEV(lg(), info) << "Requesting revert to version " << selected.version;
-
-    emit revertVersionRequested(selected);
-}
-
-}
